+++ src/stores/index.ts
@@ -2,4 +2,4 @@ export * from './modalStore';
 export * from './toastStore';
 export * from './authStore';
 export * from './filterStore';
-export * from './useRentalFormStore';
+export * from './rentalFormStore';
